backend/pyproject.toml
[project]
name = "chatbot-backend"
version = "0.1.0"
description = "Chatbot backend API with FastAPI and ChromaDB"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.9",
    "uvicorn==0.34.2",
    "python-multipart==0.0.6",
    "pydantic==2.11.5",
    "pydantic-settings==2.0.3",
    "groq==0.26.0",
    "sentence-transformers==4.1.0",
    "chromadb==1.4.0",
    "python-docx==1.1.0",
    "httpx==0.28.1",
    "python-dotenv==1.1.0",
    "fastapi-cors==0.0.6",
    "typing-extensions==4.13.2",
    "pypdf>=6.9.2",
    "pyjwt[crypto]>=2.12.1",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]
